moves files to feature folders
